complete modal mock-up with white sidebar
