+++ client/src/components/App.jsx
@@ -2,28 +2,7 @@ import React from 'react';
 import axios from 'axios';
 import Modal from 'react-modal';
 import sampleData from '../data/sampleData.js'
-
-
-const customStyles = {
-    content : {
-        top: '50%',
-        left: '50%',
-        right: 'auto',
-        bottom: 'auto',
-        marginRight: '-50%',
-        transform: 'translate(-50%, -50%)',
-        height: '625px',
-        width: '900px',
-        padding: "0",
-        border: "0",
-        background: "rgba(0, 0, 0, 0.7)",
-        overflow: "hidden"
-    },
-    overlay: {
-        background: "rgba(0, 0, 0, 0.7)",
-    },
-  };
-
+import customStyles from './style/style.js';
 
 Modal.setAppElement('#app')
 
@@ -37,25 +16,24 @@ class App extends React.Component {
             modalURL: ''
         } 
         
-        // this.getGalleryData = this.getGalleryData.bind(this);
+        this.getGalleryData = this.getGalleryData.bind(this);
         this.openModal = this.openModal.bind(this);
         // this.afterOpenModal = this.afterOpenModal.bind(this);
         this.closeModal = this.closeModal.bind(this);
     }
 
-    // componentDidMount() {
-    //     this.getGalleryData(); 
-    // }
+    componentDidMount() {
+        this.getGalleryData(); 
+    }
 
-    // getGalleryData() {
-    //     axios.get('/gallery')
-    //     .then(({data})=> this.setState({
-    //         images: data,
-    //         current: data.slice(0, 3)
-    //     }))
-    //     .catch((err)=>console.log(err))
-    // }
-
+    getGalleryData() {
+        axios.get('/gallery')
+        .then(({data})=> this.setState({
+            images: data,
+            current: data.slice(0, 3)
+        }))
+        .catch((err)=>console.log(err))
+    }
 
     openModal(e, image) {
         e.preventDefault(); 
@@ -88,13 +66,13 @@ class App extends React.Component {
           style={customStyles}
         >
           {/* <h2 ref={subtitle => this.subtitle = subtitle} style={{float: 'right'}}>Modal Mockup</h2> */}
-          <img src={this.state.modalURL} height="612px" width="900px" style={{objectFit: 'contain'}}/>
+          <img src={this.state.modalURL} height="625px" width="900px" style={{objectFit: 'contain'}}/>
+          <div style={{gridColumn: "2/span 1", backgroundColor: "white"}}>Munch Modal Mockup</div>
         </Modal>
         </div>
         )
     }
 }
-
 
 
 export default App;
