+++ client/src/components/style/style.js
@@ -1,0 +1,17 @@
+const customStyles = {
+    content : {
+        display: "grid", 
+        gridTemplateColumns: "900px 300px", 
+        height: '625px',
+        width: '1200px',
+        padding: "0",
+        border: "0",
+        background: "rgba(0, 0, 0, 0.7)",
+        overflow: "hidden"
+    },
+    overlay: {
+        background: "rgba(0, 0, 0, 0.7)",
+    },
+  };
+
+export default customStyles;
